+++ projects/angular-structured-data/src/lib/schemas/breadcrumb/breadcrumb-item.ts
@@ -1,5 +1,0 @@
-export interface BreadcrumbItem {
-  name: string;
-  url: string;
-  position: number;
-}