+++ projects/angular-structured-data/src/lib/schemas/event/attendance-mode/attendance-mode-schema.ts
@@ -1,9 +1,0 @@
-import { Schema } from '../../../schema';
-
-export abstract class AttendanceModeSchema extends Schema {
-  constructor() {
-    super(false);
-  }
-
-  abstract getEventAttendanceModeUrl(): string;
-}